requirements.txt
# AI Content Pipeline - Requirements
# Content & Growth Analyst Take-Home Assignment

# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0

# CRM Integration
hubspot-api-client>=8.3.0

# Data storage and analysis
pandas>=2.1.0
numpy>=1.24.0

# Web framework (for dashboard)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development
pytest>=7.4.0
black>=23.7.0

